Support decimal values for requests/limits.
cpu in decimal formats such as "0.5" are allowed in the kubernetes api
in addition to "1", or "100m".

"A request with a decimal point, like 0.1, is converted to 100m by the API, and precision finer than 1m is not allowed."
From: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/#requests-and-limits

+++ app/models/miq_server/worker_management/monitor/kubernetes.rb
@@ -65,8 +65,8 @@ module MiqServer::WorkerManagement::Monitor::Kubernetes
 
   def cpu_value_eql?(current, desired)
     # Convert to millicores if not already converted: "1" -> 1000; "1000m" -> 1000
-    current = current.to_s[-1] == "m" ? current.to_i : current.to_i * 1000
-    desired = desired.to_s[-1] == "m" ? desired.to_i : desired.to_i * 1000
+    current = current.to_s[-1] == "m" ? current.to_f : current.to_f * 1000
+    desired = desired.to_s[-1] == "m" ? desired.to_f : desired.to_f * 1000
     current == desired
   end
